requirements.txt
Flask==3.0.0
python-pptx==0.6.23
boto3==1.34.34
python-dotenv==1.0.0
PyPDF2==3.0.1
Werkzeug==3.0.1
